data is persistent

+++ MoodTrackerProject/MoodTracker/FileSaver.cs
@@ -9,7 +9,9 @@ public class FileSaver
     public FileSaver(string fileName)
     {
         this.fileName = fileName;
-        File.Create(this.fileName).Close();
+        if(!File.Exists(this.fileName)) {
+            File.Create(this.fileName).Close();
+        }
     }
 
     public void AppendLine(string line) {
